pyproject.toml
[tool.poetry]
name = "mantos"
version = "0.1.0"
description = ""
authors = ["Lucas Ferreira <lucasferreiraek@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.*"
fastapi = "^0.108.0"
uvicorn = "^0.25.0"
pydantic-settings = "^2.1.0"
sqlalchemy = "^2.0.23"
alembic = "^1.13.1"
psycopg2-binary = "^2.9.9"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
taskipy = "^1.12.2"
blue = "^0.9.1"
ruff = "^0.1.9"
httpx = "^0.26.0"
isort = "^5.13.2"
factory-boy = "^3.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 79
exclude = ['.venv', 'migrations']

[tool.blue]
extend-exclude = '(migrations/)'

[tool.taskipy.tasks]
lint = 'ruff . && blue --check . --diff'
format = 'blue .  && isort .'
run = 'uvicorn mantos.app:app --reload'
pre-test = 'task lint'
test = 'pytest -s -x --cov=mantos -vv'
post-test = 'coverage html'
